[update] examples/lang_{json, pl0, xml} の構造を lang_easyarith に合わせた
