+++ README.md
@@ -39,12 +39,12 @@ copager = { ..., features = ["derive", "regexlex", "lalr1", "void"] }
 
 ## Examples
 
-- [example_build_oneshot](examples/build_oneshot)
-- [example_build_prebuild](examples/build_prebuild)
+- [example_build_oneshot](examples/build_oneshot) [(main.rs)](examples/build_oneshot/src/main.rs)
+- [example_build_prebuild](examples/build_prebuild) [(main.rs)](examples/build_prebuild/src/main.rs)
 - [example_lang_easyarith](examples/lang_easyarith) [(syntax.rs)](examples/lang_easyarith/src/syntax.rs)
-- [example_lang_json](examples/lang_json) [(lib.rs)](examples/lang_json/src/lib.rs)
-- [example_lang_pl0](examples/lang_pl0) [(lib.rs)](examples/lang_pl0/src/lib.rs)
-- [example_lang_xml](examples/lang_xml) [(lib.rs)](examples/lang_xml/src/lib.rs)
+- [example_lang_json](examples/lang_json) [(syntax.rs)](examples/lang_json/src/syntax.rs)
+- [example_lang_pl0](examples/lang_pl0) [(syntax.rs)](examples/lang_pl0/src/syntax.rs)
+- [example_lang_xml](examples/lang_xml) [(syntax.rs)](examples/lang_xml/src/syntax.rs)
 
 ```
 $ cargo run -p example_lang_easyarith
