+++ examples/lang_json/src/main.rs
@@ -4,7 +4,7 @@ use copager::template::LALR1;
 use copager::ir::SExp;
 use copager::Processor;
 
-use example_lang_json::Json;
+use example_lang_json::syntax::Json;
 
 type Config = LALR1<Json>;
 type MyProcessor = Processor<Config>;
