+++ examples/lang_pl0/src/main.rs
@@ -4,7 +4,7 @@ use copager::template::LALR1;
 use copager::ir::SExp;
 use copager::Processor;
 
-use example_lang_pl0::Pl0;
+use example_lang_pl0::syntax::Pl0;
 
 type Config = LALR1<Pl0>;
 type MyProcessor = Processor<Config>;
