+++ examples/lang_xml/src/main.rs
@@ -4,7 +4,7 @@ use copager::template::LALR1;
 use copager::ir::SExp;
 use copager::Processor;
 
-use example_lang_xml::Xml;
+use example_lang_xml::syntax::Xml;
 
 type Config = LALR1<Xml>;
 type MyProcessor = Processor<Config>;
